+++ packages/react-resizable-panels-website/tests/Panel-OnCollapse.spec.ts
@@ -28,18 +28,38 @@ async function verifyEntries(
 
 test.describe("Panel onCollapse prop", () => {
   test.beforeEach(async ({ page }) => {
-    await page.goto("http://localhost:2345/examples/imperative-api");
+    await page.goto("http://localhost:2345/examples/imperative-api?onCollapse");
   });
 
   test("should be called once on-mount", async ({ page }) => {
     // No panels are collapsed by default.
-    await verifyEntries(page, []);
+    await verifyEntries(page, [
+      { panelId: "left", collapsed: false },
+      { panelId: "middle", collapsed: false },
+      { panelId: "right", collapsed: false },
+    ]);
 
-    // If we override via URL parameters, we should be notified though.
-    await page.goto("http://localhost:2345/examples/imperative-api?collapse");
+    // If we override via URL parameters, left and right panels should be collapsed by default.
+    await page.goto(
+      "http://localhost:2345/examples/imperative-api?collapse&onCollapse"
+    );
     await verifyEntries(page, [
       { panelId: "left", collapsed: true },
+      { panelId: "middle", collapsed: false },
       { panelId: "right", collapsed: true },
+    ]);
+  });
+
+  // Edge case
+  test("should only call onCollapse for panels that are collapsible", async ({
+    page,
+  }) => {
+    await page.goto(
+      "http://localhost:2345/examples/imperative-api?noMiddleCollapse&onCollapse"
+    );
+    await verifyEntries(page, [
+      { panelId: "left", collapsed: false },
+      { panelId: "right", collapsed: false },
     ]);
   });